requirements.txt
blinker==1.9.0
click==8.3.0
et_xmlfile==2.0.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
